try-on-service/requirements.txt
# Try-On Service - Virtual Try-On API
# FastAPI service using OOTDiffusion for virtual clothing try-on

fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow==10.2.0
numpy==1.26.3
aiofiles==23.2.1

# AI/ML Dependencies
torch==2.2.0
torchvision==0.17.0
transformers==4.37.2
diffusers==0.25.1
accelerate==0.26.1
safetensors==0.4.2

# Image processing
opencv-python-headless==4.9.0.80
rembg==2.0.50

# Utilities
httpx==0.26.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
